requirements.txt
hydra-core==1.1.1
joblib==1.1.0
matplotlib==3.5.1
numpy==1.22.3
omegaconf==2.1.1
pandas==1.4.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
tables==3.7.0
tensorflow==2.8.0
